shopKeeperRotateFn_801e7c4c: drop redundant (int) in (s16)(int) yaw store
(s16)(int)(floatExpr) sign-extended a needless int intermediate; (s16)
stores the fctiwz result directly. 96.58->97.63%.

+++ src/main/dll/SP/dll_0286_spshopkeeper.c
@@ -448,7 +448,7 @@ f32 shopKeeperRotateFn_801e7c4c(s16* obj, void* player, int mode)
             {
                 diff = 0;
             }
-            *obj = (s16)(int)((f32)(diff >> 3) * timeDelta + (f32) * obj);
+            *obj = (s16)((f32)(diff >> 3) * timeDelta + (f32) * obj);
         }
     }
     return dist;
